Разбиваем на классы, базовый уже нечитаем
